feat: S3 이미지 업로드, 페이징 기능 구현

+++ src/main/java/com/example/board/post/controller/PostController.java
@@ -1,12 +1,11 @@
 package com.example.board.post.controller;
 
-import com.example.board.board.entity.Board;
 import com.example.board.post.dto.request.PostEditRequest;
-import com.example.board.post.dto.request.PostOneRequest;
+import com.example.board.post.dto.request.PostOneResponse;
 import com.example.board.post.dto.request.PostRequest;
-import com.example.board.post.dto.response.PostResponse;
 import com.example.board.post.entity.Post;
 import com.example.board.post.service.PostService;
+import java.net.URLDecoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -25,45 +24,41 @@ import org.springframework.web.bind.annotation.RestController;
 public class PostController {
     private final PostService postService;
 
-    // 현재 사용자 정체
-//    @GetMapping("/hi")
-//    public void hi(@PathVariable(name = "boardCategory") String boardCategory,
-//                   @RequestHeader("X-Authorization-email") String email,
-//                   @RequestHeader("X-Authorization-nickname") String nickname) {
-//        System.out.println(boardCategory);
-//        System.out.println(email);
-//        System.out.println(nickname);
-//    }
-
     // 게시글 조회
     @GetMapping("/{postId}")
-    public PostOneRequest getPost(@PathVariable(name = "postId") Long postId) {
+    public PostOneResponse getPost(@PathVariable(name = "postId") Long postId) {
         return postService.getPost(postId);
     }
 
     // 게시글 생성
     @PostMapping
-    public PostOneRequest addPost(@PathVariable(name = "boardCategory") String boardCategory,
-                                  @RequestBody PostRequest postRequest,
-                                  @RequestHeader("X-Authorization-email") String email,
-                                  @RequestHeader("X-Authorization-nickname") String nickname,
-                                  @RequestHeader("X-Authorization-memberId") Long memberId) {
-        System.out.println(email);
-        System.out.println(nickname);
-        System.out.println(memberId);
+    public PostOneResponse addPost(@PathVariable(name = "boardCategory") String boardCategory,
+                                   @RequestBody PostRequest postRequest,
+                                   @RequestHeader("X-Authorization-email") String email,
+                                   @RequestHeader("X-Authorization-nickname") String nickname,
+                                   @RequestHeader("X-Authorization-memberId") Long memberId) {
+
+        nickname = URLDecoder.decode(nickname);
         return postService.savePost(boardCategory, postRequest, nickname, memberId);
     }
 
+    // 게시글 삭제
     @DeleteMapping("/{postId}")
     public ResponseEntity<String> deletePost(@PathVariable(name = "postId") Long postId) {
+        if (postId == null) {
+            return ResponseEntity.badRequest().body("postId가 유효 하지 않습니다.");
+        }
         postService.deletePost(postId);
         return ResponseEntity.ok("게시글이 삭제되었습니다.");
     }
 
-//    @PutMapping("/{postid}")
-//    public ResponseEntity<Post> updatePost(@PathVariable(name = "postid") Long postId,
-//                                           @RequestBody PostEditRequest postEditRequest) {
-//        Post updatedPost = postService.updatePost(postId, postEditRequest);
-//        return ResponseEntity.ok(updatedPost);
-//    }
+    // 게시글 수정
+    @PutMapping("/{postId}")
+    public ResponseEntity<Post> updatePost(@PathVariable(name = "postId") Long postId,
+                                           @RequestBody PostEditRequest postEditRequest) {
+
+        Post updatedPost = postService.updatePost(postId, postEditRequest);
+        return ResponseEntity.ok(updatedPost);
+    }
+
 }
